setup: tool.mysql2pgsql update

+++ nix/tools/mysql2pgsql.nix
@@ -1,4 +1,4 @@
-# generated using pypi2nix tool (version: 1.5.0.dev0)
+# generated using pypi2nix tool (version: 1.8.0)
 # See more at: https://github.com/garbas/pypi2nix
 #
 # COMMAND:
@@ -13,12 +13,11 @@ let
   inherit (pkgs) makeWrapper;
   inherit (pkgs.stdenv.lib) fix' extends inNixShell;
 
-  python27Full = pkgs.python27.override { x11Support=true; };
-  pythonPackages = import "${toString pkgs.path}/pkgs/top-level/python-packages.nix" {
+  pythonPackages =
+  import "${toString pkgs.path}/pkgs/top-level/python-packages.nix" {
     inherit pkgs;
     inherit (pkgs) stdenv;
-    python = python27Full;
-    self = pythonPackages;
+    python = pkgs.python27Full;
   };
 
   commonBuildInputs = with pkgs; [ postgresql mysql.lib ];
@@ -32,11 +31,13 @@ let
         buildInputs = [ makeWrapper ] ++ (builtins.attrValues pkgs);
         buildCommand = ''
           mkdir -p $out/bin
-          ln -s ${pythonPackages.python.interpreter} $out/bin/${pythonPackages.python.executable}
-          for dep in ${builtins.concatStringsSep " " (builtins.attrValues pkgs)}; do
+          ln -s ${pythonPackages.python.interpreter}               $out/bin/${pythonPackages.python.executable}
+          for dep in ${builtins.concatStringsSep " "               (builtins.attrValues pkgs)}; do
             if [ -d "$dep/bin" ]; then
               for prog in "$dep/bin/"*; do
-                ln -s $prog $out/bin/`basename $prog`
+                if [ -f $prog ]; then
+                  ln -s $prog $out/bin/`basename $prog`
+                fi
               done
             fi
           done
@@ -62,7 +63,103 @@ let
 
   python = withPackages {};
 
-  generated = import ./mysql2pgsql_generated.nix { inherit pkgs python commonBuildInputs commonDoCheck; };
+  generated = self: {
+
+    "MySQL-python" = python.mkDerivation {
+      name = "MySQL-python-1.2.5";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/a5/e9/51b544da85a36a68debe7a7091f068d802fc515a3a202652828c73453cad/MySQL-python-1.2.5.zip"; sha256 = "811040b647e5d5686f84db415efd697e6250008b112b6909ba77ac059e140c74"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.gpl1;
+        description = "Python interface to MySQL";
+      };
+    };
+
+
+
+    "PyYAML" = python.mkDerivation {
+      name = "PyYAML-3.12";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/4a/85/db5a2df477072b2902b0eb892feb37d88ac635d36245a72a6a69b23b383a/PyYAML-3.12.tar.gz"; sha256 = "592766c6303207a20efc445587778322d7f73b161bd994f227adaa341ba212ab"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.mit;
+        description = "YAML parser and emitter for Python";
+      };
+    };
+
+
+
+    "argparse" = python.mkDerivation {
+      name = "argparse-1.4.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/18/dd/e617cfc3f6210ae183374cd9f6a26b20514bbb5a792af97949c5aacddf0f/argparse-1.4.0.tar.gz"; sha256 = "62b089a55be1d8949cd2bc7e0df0bddb9e028faefc8c32038cc84862aefdd6e4"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.psfl;
+        description = "Python command-line parsing library";
+      };
+    };
+
+
+
+    "psycopg2" = python.mkDerivation {
+      name = "psycopg2-2.7";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/62/ca/0a479c9664526e86c2913a7ad593586eeb86b428b7e629e7c7b6b69e3cb7/psycopg2-2.7.tar.gz"; sha256 = "ceadecf660ad4f7a31ea5baef30a7351add8626f9fd3daaafabb9a9e549f3f9a"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.lgpl2;
+        description = "psycopg2 - Python-PostgreSQL Database Adapter";
+      };
+    };
+
+
+
+    "py-mysql2pgsql" = python.mkDerivation {
+      name = "py-mysql2pgsql-0.1.5";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/28/49/c82e4313c63b26224950bef47ab19f52a921dd4c279abf31aa3ee9d9cbbe/py-mysql2pgsql-0.1.5.tar.gz"; sha256 = "024d3f1d3994679421102a2cb70c1fde1306bdf8c70839db907a728324c65cdf"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [
+      self."MySQL-python"
+      self."PyYAML"
+      self."argparse"
+      self."psycopg2"
+      self."termcolor"
+    ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.mit;
+        description = "Tool for migrating/converting from mysql to postgresql.";
+      };
+    };
+
+
+
+    "termcolor" = python.mkDerivation {
+      name = "termcolor-1.1.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/8a/48/a76be51647d0eb9f10e2a4511bf3ffb8cc1e6b14e9e4fab46173aa79f981/termcolor-1.1.0.tar.gz"; sha256 = "1d6d69ce66211143803fbc56652b41d73b4a400a2891d7bf7a1cdf4c02de613b"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.mit;
+        description = "ANSII Color formatting for output in terminal.";
+      };
+    };
+
+  };
   overrides = import ./mysql2pgsql_override.nix { inherit pkgs python; };
 
 in python.withPackages (fix' (extends overrides generated))
